requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0

# OpenAI (LLM only)
openai>=1.0.0

# Audio Recording & Processing
sounddevice>=0.4.6
soundfile>=0.12.1
webrtcvad-wheels>=2.0.10
numpy>=1.24.0

# STT - Local Whisper (FREE)
openai-whisper>=20231117

# TTS - Edge TTS (FREE)
edge-tts>=6.1.0

# Wake Word Detection
pvporcupine>=3.0.0

# Audio Playback
pygame>=2.5.0

# Web Dashboard
flask>=3.0.0
flask-socketio>=5.3.0

# Utilities
colorlog>=6.7.0
